+++ lib/value/Cargo.toml
@@ -22,12 +22,14 @@ mlua = { version = "0.7.3", default-features = false, features = ["lua54", "send
 serde = { version = "1.0.136", default-features = false, features = ["derive", "rc"], optional = true }
 serde_json = { version = "1.0.78", default-features = false, optional = true }
 toml = { version = "0.5.8", default-features = false, optional = true }
+quickcheck = {version = "1.0.3", optional = true}
 
 [features]
 lua = ["mlua"]
 api = ["async-graphql"]
 json = ["serde", "serde_json"]
 test = []
+arbitrary = ["quickcheck"]
 
 [dev-dependencies]
 quickcheck = "1.0.3"
