+++ lib/vector-core/src/event/test/common.rs
@@ -3,7 +3,6 @@ use std::{
     iter,
 };
 
-use bytes::Bytes;
 use chrono::{DateTime, NaiveDateTime, Utc};
 use quickcheck::{empty_shrinker, Arbitrary, Gen};
 
@@ -13,7 +12,6 @@ use crate::event::{
 };
 
 const MAX_F64_SIZE: f64 = 1_000_000.0;
-const MAX_ARRAY_SIZE: usize = 4;
 const MAX_MAP_SIZE: usize = 4;
 const MAX_STR_SIZE: usize = 16;
 const ALPHABET: [&str; 27] = [
@@ -546,42 +544,6 @@ impl Arbitrary for MetricData {
                     })
                 }),
         )
-    }
-}
-
-impl Arbitrary for Value {
-    fn arbitrary(g: &mut Gen) -> Self {
-        // Quickcheck can't derive Arbitrary for enums, see
-        // https://github.com/BurntSushi/quickcheck/issues/98.  The magical
-        // constant here are the number of fields in `Value`. Because the field
-        // total is a power of two we, happily, don't introduce a bias into the
-        // field picking.
-        match u8::arbitrary(g) % 8 {
-            0 => {
-                let bytes: Vec<u8> = Vec::arbitrary(g);
-                Value::Bytes(Bytes::from(bytes))
-            }
-            1 => Value::Integer(i64::arbitrary(g)),
-            2 => {
-                let mut f = f64::arbitrary(g) % MAX_F64_SIZE;
-                if f.is_nan() {
-                    f = 0.0;
-                }
-                Value::from(f)
-            }
-            3 => Value::Boolean(bool::arbitrary(g)),
-            4 => Value::Timestamp(datetime(g)),
-            5 => {
-                let mut gen = Gen::new(MAX_MAP_SIZE);
-                Value::Map(BTreeMap::arbitrary(&mut gen))
-            }
-            6 => {
-                let mut gen = Gen::new(MAX_ARRAY_SIZE);
-                Value::Array(Vec::arbitrary(&mut gen))
-            }
-            7 => Value::Null,
-            _ => unreachable!(),
-        }
     }
 }
